Refactor main package
create internal package and move implementation into this package
