+++ tooling/image-sync/internal/repository_test.go
@@ -1,4 +1,4 @@
-package main
+package internal
 
 import (
 	"context"
